requirements.txt
aiodns==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
ccxt==3.0.59
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.12.4
Flask==3.0.0
fonttools==4.43.1
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.10.0
gym==0.26.2
gym-notices==0.0.8
gym-trading-env==0.3.3
gymnasium==0.29.1
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
nest-asyncio==1.5.8
networkx==3.2
numpy==1.26.1
packaging==23.2
pandas==2.1.1
peewee==3.17.0
Pillow==10.1.0
prettytable==3.9.0
pycares==4.4.0
pycparser==2.21
pyecharts==2.0.4
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
stable-baselines3==2.1.0
sympy==1.12
torch==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
webencodings==0.5.1
Werkzeug==3.0.1
yarl==1.9.2
yfinance==0.2.31
alpaca-py==0.13.1